Move input component into common folder
